refactor: Rename Pair to PairScreen and move AmountValueHeader to components
