+++ lib/components/wallet_connect/components/amount_value_header.dart
