+++ lib/pages/main/wallet_contents/add_wallet_connect/add_wallet_connect.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:mobile_scanner/mobile_scanner.dart';
-import 'package:qubic_wallet/components/wallet_connect/pair.dart';
+import 'package:qubic_wallet/components/wallet_connect/pair_screen.dart';
 import 'package:qubic_wallet/config.dart';
 import 'package:qubic_wallet/di.dart';
 import 'package:qubic_wallet/flutter_flow/theme_paddings.dart';
@@ -127,7 +127,7 @@ class _AddWalletConnectState extends State<AddWalletConnect> {
           bool? userhasConfirmed =
               await Navigator.of(context).push(MaterialPageRoute<bool>(
                   builder: (BuildContext context) {
-                    return Pair(
+                    return PairScreen(
                       pairingId: wcPairingId!,
                       pairingEvents: wcPairingEvents,
                       pairingMethods: wcPairingMethods,
